feat!: Port to Next.js
Ports website to Next.js (https://nextjs.org/). This allows dynamic
components and css frameworks in the future. Font distribution and
custom pages are not supported yet. The website is rendered into
static html for production.

This also runs CI to lint and test builds on commit pushes to any
branch. Prettier is also run by default for javascript and markdown
files.