Add Link and InternalLink components
Link wraps <a> with custom styles, it should be used for most links.

InternalLink wraps "next/link", it should only be used for links to
internal pages that need React/Next routing.

For both of these components, only `href` and child elements are
used. `className` is appended to the wrapped <a>.